feat: Add Farcaster user and on-chain balance to views
This commit introduces the ability to display Farcaster user information and the user's on-chain balance within various game views. It also adds a click handler for the account display.

The following views have been updated to accept and pass down these new props:
- StatusBoard
- PartyView
- DepartView
- RecoveryView
- GachaView

This change lays the groundwork for deeper integration with Farcaster and on-chain wallet functionalities within the Mining Quest application.